backend/requirements.txt
# =============================================================================
# GraphRAG Chatbot Platform - Backend Dependencies
# =============================================================================

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
sse-starlette==2.0.0

# Database - PostgreSQL
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Database - Neo4j
neo4j==5.17.0

# Database - Qdrant (Vector Store)
qdrant-client==1.7.3

# Cache & Message Queue
redis==5.0.1
celery==5.3.6
flower==2.0.1

# LLM & AI
langchain==0.2.16
langchain-community==0.2.16
langchain-ollama==0.1.3
langchain-core==0.2.40

# PDF Processing
pdfplumber==0.10.4
pypdf==4.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
tenacity==8.2.3
aiofiles==23.2.1

# =============================================================================
# Development Dependencies
# =============================================================================
# pytest==8.0.0
# pytest-asyncio==0.23.4
# pytest-cov==4.1.0
# httpx==0.27.0
# black==24.1.1
# ruff==0.2.0
# mypy==1.8.0
